'tone' --> 'letter'
Changed Tone class to Letter class, and any value with 'tone(s)' in the name to 'letter(s)'. The concept being conveyed in these values is the 'letter name,' and is used in conjunction with the 'accidental' name. Calling these 'letter' values is more clear than calling them 'tone' values.